requirements.txt
bittensor>=6.0.0
pyyaml>=6.0.1
inquirer>=3.1.3
rich>=13.7.0
flask>=3.0.0
flask-wtf>=1.2.1
python-dotenv>=1.0.0
plotext>=5.2.8 